requirements.txt
Pillow
numpy
sklearn
colorsys
